updated READ, corrected all edgecases for currently applied functionalities, added % escape

+++ include/ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: jormanue <jormanue@student.42porto.co      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/04/30 16:53:59 by jormanue          #+#    #+#             */
-/*   Updated: 2025/05/20 19:57:01 by jormanue         ###   ########.fr       */
+/*   Updated: 2025/05/21 14:38:43 by jormanue         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -19,7 +19,7 @@ int		ft_printf(char const *s, ...);
 int		ft_putchar(char c);
 int		ft_printstr(char *s);
 size_t	ft_strlen(const char *s);
-int		ft_putnbr(int nb);
+int		ft_putnbr(long nb);
 int		ft_seetype(unsigned char *point, va_list param);
 
 #endif
